style(module): extracted the tests module from static module
